# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/admin/configs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all configs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Configs"
                ],
                "summary": "List all configs",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.ListAllConfigsRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new config with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Configs"
                ],
                "summary": "Create a new config",
                "parameters": [
                    {
                        "description": "Config to create",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.CreateConfigParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.CreateConfigRow"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/configs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a config by its ID",
                "tags": [
                    "Admin Configs"
                ],
                "summary": "Get a config by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.GetConfigByIDRow"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update config details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Configs"
                ],
                "summary": "Update an existing config",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated config details",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.UpdateConfigParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.UpdateConfigRow"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a config by its ID",
                "tags": [
                    "Admin Configs"
                ],
                "summary": "Delete a config by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "go-wordpress_internal_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "go-wordpress_internal_storage_sql_sqlc.CreateConfigParams": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "websiteID": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
        "go-wordpress_internal_storage_sql_sqlc.CreateConfigRow": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "key": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "websiteID": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
        "go-wordpress_internal_storage_sql_sqlc.EntityStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive",
                "in_progress"
            ],
            "x-enum-varnames": [
                "EntityStatusActive",
                "EntityStatusInactive",
                "EntityStatusInProgress"
            ]
        },
        "go-wordpress_internal_storage_sql_sqlc.GetConfigByIDRow": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "key": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "websiteID": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
        "go-wordpress_internal_storage_sql_sqlc.ListAllConfigsRow": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "key": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "websiteID": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
        "go-wordpress_internal_storage_sql_sqlc.UpdateConfigParams": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "key": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.EntityStatus"
                },
                "value": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "websiteID": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
        "go-wordpress_internal_storage_sql_sqlc.UpdateConfigRow": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "key": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "websiteID": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token in the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}